pyproject.toml
[tool.poetry]
name = "nonebot-plugin-quote"
version = "0.2.3"
description = "一款适用于QQ群聊天的语录库插件"
authors = ["RongRongJi <316315867@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_quote"}]

[tool.poetry.dependencies]
python = "^3.8"
jieba = "^0.42.1"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = "^2.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
